// libs/shared/src/domain/models/stocks/crypto-global-d1.model.ts
export interface CryptoGlobalD1 {
  id: number;
  Symbol: string;
  OpenTime: Date;
  Open: number;
  High: number;
  Low: number;
  Close: number;
  Volume: number;
  CloseTime: Date;
  QuoteAssetVolume: number;
  NumberOfTrades: number;
  TakerBuyBaseAssetVolume: number;
  TakerBuyQuoteAssetVolume: number;
}